Reorganized assets folder & updated read me
