+++ README.md
@@ -1,1 +1,21 @@
-## ReadMe draft in progress
+# Brian Robert Cajulis's Portfolio Website
+Hello! Welcome to my portfolio website.  I developed this website using React, Bootstrap, and a few NPM packages.  If you have any comments on what I can improve, feel free to leave it in the issues section of this repository.
+
+[**# Wireframing**](#Wireframing)
+
+## Wireframing
+
+[# Screenshots](#screenshots) - Wireframe Overview
+[# Conclusion](#conclusion) - Final Thoughts
+
+Before proceeding to develop a project, I always like to get the basic layout of what it will look like.  The platform I used to create these wireframes is [Lucidchard](https://www.lucidchart.com/).
+
+### Screenshots
+Want a close up view of mockup 1? [Click Here](https://www.lucidchart.com/invitations/accept/ad592b7f-ea48-4cee-856f-f0fc1e30ecd3)
+![Alt text](src/assets/screenshots/portfolio-mockup-1?raw=true"Portfolio Mockup 1")
+
+Want a close up view of mockup 2? [Click Here](https://www.lucidchart.com/invitations/accept/8d9e6556-2c9a-411e-af50-3f944ca5c314)
+![Alt text](src/assets/screenshots/portfolio-mockup-2?raw=true"Portfolio Mockup 1")
+
+### Conclusion
+Developing the second version of my portfolio website has given me a lot of insight about how much I've improved over the years.  It also reminded me of a lesson I learned while studying at Azusa Pacific University.  Plans change over the course of development.  I may have chosen to do one thing, but then decide something else was better.  The most important skill to exercise is to adapt and learn.  I will continue to update this website until I decide to create new version.  The feedback I've received from friends and colleagues helped me make many improvements.
