--- README.old.md
@@ -1,2 +1,0 @@
-# portfolio-website-v2
-[Current, Created 2018] Shows my recent development work, education history, and side projects.  This website is developed using the React framework and JavaScript (ES6+)
