+++ src/assets/images/bootstrap.png
