+++ src/containers/about/About.scss
@@ -1,10 +1,10 @@
 .about {
     background: 
         linear-gradient(
-        rgba(0, 19, 18, 0.897), 
-        rgba(0, 95, 90, 0.89), 
+        rgba(0, 19, 18, 0.815), 
+        rgba(0, 95, 90, 0.733), 
         ),
-        url('../../assets/about-background.jpg');
+        url('../../assets/images/about-background.jpg');
     background-size: cover;
     p {
         font-size: 4vw;
